Refactor project to MVC structure
