+++ src/routes/__init__.py
@@ -1,35 +1,29 @@
-# acapella_backend/main.py
-
-from flask import Flask, request, jsonify
-from flask_cors import CORS
-import subprocess
+from flask import Blueprint, request, jsonify
 import os
-import requests
-import yt_dlp
 import uuid
+import yt_dlp
+import requests
 from bs4 import BeautifulSoup
+from src.utils.lalalai_splitter import batch_process
 
-app = Flask(__name__)
-CORS(app)
-app.secret_key = 'supersecretkey'  # Replace with env-secured key
+routes = Blueprint('routes', __name__)
 
-@app.route('/process', methods=['POST'])
+@routes.route('/process', methods=['POST'])
 def process_song():
     data = request.get_json()
     youtube_url = data.get('url')
-    lalal_license = os.getenv("lalalal_ai_license")
+    lalal_license = os.getenv('lalalal_ai_license')
     if not youtube_url or not lalal_license:
         return jsonify({'error': 'URL and license required'}), 400
 
     filename = f"./downloads/{uuid.uuid4()}.mp3"
-
     try:
         ydl_opts = {
             'format': 'bestaudio/best',
             'noplaylist': True,
             'outtmpl': filename,
             'quiet': True,
-            'ffmpeg_location': "C:/Users/galga/Downloads/ffmpeg-2025-06-04-git-a4c1a5b084-essentials_build/ffmpeg-2025-06-04-git-a4c1a5b084-essentials_build/bin",  # ✅ Update this
+            'ffmpeg_location': "C:/Users/galga/Downloads/ffmpeg-2025-06-04-git-a4c1a5b084-essentials_build/ffmpeg-2025-06-04-git-a4c1a5b084-essentials_build/bin",
             'postprocessors': [{
                 'key': 'FFmpegExtractAudio',
                 'preferredcodec': 'mp3'
@@ -41,24 +35,15 @@ def process_song():
         return jsonify({'error': f'Failed to download audio: {str(e)}'}), 500
 
     try:
-        output_dir = os.path.join("acapellas", os.path.splitext(os.path.basename(filename))[0])
+        output_dir = os.path.join('acapellas', os.path.splitext(os.path.basename(filename))[0])
         os.makedirs(output_dir, exist_ok=True)
-        subprocess.run([
-            'python', 'lalalai_splitter.py',
-            '--license', lalal_license,
-            '--input', filename + ".mp3",
-            '--output', output_dir,
-            '--splitter', 'orion',
-            '--stem', 'vocals',
-            '--enhanced-processing', 'False',
-            '--noise-cancelling', '1'
-        ], check=True)
-    except subprocess.CalledProcessError as e:
+        batch_process(lalal_license, filename + '.mp3', output_dir, 'vocals', 'orion', False, 1)
+    except Exception:
         return jsonify({'error': 'Lalal.ai processing failed'}), 500
 
     return jsonify({'message': 'Success', 'path': output_dir})
 
-@app.route('/playlist', methods=['GET'])
+@routes.route('/playlist', methods=['GET'])
 def get_playlist():
     playlist_url = request.args.get('url')
     if not playlist_url:
@@ -69,7 +54,6 @@ def get_playlist():
         'extract_flat': True,
         'skip_download': True
     }
-
     try:
         with yt_dlp.YoutubeDL(ydl_opts) as ydl:
             info = ydl.extract_info(playlist_url, download=False)
@@ -80,13 +64,11 @@ def get_playlist():
                     'id': entry.get('id'),
                     'url': f'https://www.youtube.com/watch?v={entry.get("id")}'
                 })
-
             return jsonify({'playlist_title': info.get('title'), 'results': videos})
-
     except Exception as e:
         return jsonify({'error': str(e)}), 500
 
-@app.route('/search_playlists_html', methods=['GET'])
+@routes.route('/search_playlists_html', methods=['GET'])
 def search_playlists_html():
     query = request.args.get('q')
     if not query:
@@ -94,18 +76,15 @@ def search_playlists_html():
 
     search_url = f"https://www.youtube.com/results?search_query={query.replace(' ', '+')}+playlist&sp=EgIQAw%253D%253D"
     headers = {'User-Agent': 'Mozilla/5.0'}
-
     try:
         r = requests.get(search_url, headers=headers)
         soup = BeautifulSoup(r.text, 'html.parser')
-
-        # Not perfect — YouTube uses JS for most content — we’d need yt-dlp or Selenium for full results
         titles = [tag.text for tag in soup.select('a#video-title')]
         return jsonify({'results': titles[:5]})
     except Exception as e:
         return jsonify({'error': str(e)}), 500
 
-@app.route('/search_playlists', methods=['GET'])
+@routes.route('/search_playlists', methods=['GET'])
 def search_playlists():
     query = request.args.get('q')
     if not query:
@@ -115,16 +94,13 @@ def search_playlists():
         'quiet': True,
         'extract_flat': False,
         'skip_download': True,
-        'default_search': f"ytsearchdate5:{query}",  # Prioritize newer results
+        'default_search': f"ytsearchdate5:{query}"
     }
-
     playlists = []
     try:
         with yt_dlp.YoutubeDL(ydl_opts) as ydl:
             info = ydl.extract_info(query, download=False)
-
             entries = info.get('entries', [])
-            print(entries)
             for entry in entries:
                 if entry.get('_type') == 'playlist':
                     playlists.append({
@@ -132,17 +108,13 @@ def search_playlists():
                         'uploader': entry.get('uploader'),
                         'playlist_url': f"https://www.youtube.com/playlist?list={entry.get('id')}"
                     })
-
         if not playlists:
             return jsonify({'message': 'No playlists found.', 'results': []}), 200
-
         return jsonify({'results': playlists})
-
     except Exception as e:
         return jsonify({'error': str(e)}), 500
 
-
-@app.route('/search', methods=['GET'])
+@routes.route('/search', methods=['GET'])
 def search_youtube():
     query = request.args.get('q')
     if not query:
@@ -154,7 +126,6 @@ def search_youtube():
         'extract_flat': False,
         'default_search': 'ytsearch5'
     }
-
     results = []
     try:
         with yt_dlp.YoutubeDL(ydl_opts) as ydl:
@@ -167,10 +138,7 @@ def search_youtube():
                 })
     except Exception as e:
         return jsonify({'error': str(e)}), 500
-
     return jsonify({'results': results})
 
-if __name__ == '__main__':
-    os.makedirs('downloads', exist_ok=True)
-    os.makedirs('acapellas', exist_ok=True)
-    app.run(debug=True)
+def register_routes(app):
+    app.register_blueprint(routes)
